Insert title, date, previous and next into articles
Using the information from the "head" elements of all the articles,
go through all the articles adding that information to elements that
have been tagged with the required class names.

For example, there is no need to repeat the title in the "title"
element and the h1 element. And the creation dates of the articles
can be used to populate links to the next and previous articles in
each thread.